<refactor> barrel files utils
