+++ frontend/src/app/pages/empregado/solicitacao/listar-atribuicao/listar-atribuicao.component.ts
@@ -5,7 +5,7 @@ import { CommonModule } from "@angular/common";
 import { RouterModule } from "@angular/router";
 import { EstadoAmigavelPipe } from "../../../../shared/pipes/estado-amigavel.pipe";
 import { EstadoCorPipe } from "../../../../shared/pipes/estado-cor.pipe";
-import { HistoricoUtils } from "../../../../shared/utils/historico-utils";
+import { HistoricoUtils } from "../../../../shared/utils";
 
 @Component({
   selector: "app-listar-atribuicao",
