+++ frontend/src/app/shared/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './estado-amigavel.pipe';
+export * from './estado-cor.pipe';
