+++ frontend/src/app/shared/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './data-utils';
+export * from './endereco-utils';
+export * from './historico-utils';
+export * from './number-utils';
